iai_boxy_base: improved mesh of the base
Moved the lasers to their real position.
Fixed the laser axes.
Added the bars for the head mounting.
Added the plate on top of the bars.
The black adapter plate (6mm) is still missing below the PTU.

Now you can clearly see that the head is not placed correctly in the URDF.
This will need fixing, but I have to be careful because it affects
the hand/eye calibration.